+++ doc/user/Changelog.md
@@ -47,6 +47,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Fix coordinate fallback when linking locations in the street graph [#7056](https://github.com/opentripplanner/OpenTripPlanner/pull/7056)
 - Handle barrier entrances correctly [#7067](https://github.com/opentripplanner/OpenTripPlanner/pull/7067)
 - Fix default via vertex name translation mapping [#7046](https://github.com/opentripplanner/OpenTripPlanner/pull/7046)
+- To fix arriveBy searches always set arriveBy property to false in transfer requests [#7089](https://github.com/opentripplanner/OpenTripPlanner/pull/7089)
 [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.8.1 (2025-09-22)